remove hard-coded version string from req header
Use the gem version constant instead of the string existing in
multiple places. This will prevent confusion in case the client
version is bumped but the request header gets missed. This could lead
to potential confusion if Bitpay ever needs to debug version of Ruby
client used to access the API.