Rakefile to clear added api keys
Local tests are run many times a day, but the Travis.CI tests are run
against test server accounts. So the default task (for Travis.CI) does
not clean up all API keys from the local server, but there is a rake
task that can be run locally or remotely to do this for you.